requirements.txt
absl-py==0.4.0
aiodns==1.1.1
aiohttp==2.3.7
alabaster==0.7.9
#anaconda-client==1.6.0
#anaconda-project==0.4.1
asciichart==0.1
astor==0.7.1
astroid==1.4.9
astropy==2.0.3
async-timeout==2.0.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
backtrader==1.9.64.122
beautifulsoup4==4.5.3
bitarray==0.8.1
#bitmex-easy-data-scripts==0.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.4
boto==2.48.0
boto3==1.6.22
botocore==1.9.22
Bottleneck==1.2.1
bz2file==0.98
cchardet==2.1.1
ccxt==1.10.543
certifi==2018.8.13
cffi==1.9.1
chardet==3.0.4
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
#comtypes==1.1.2
#conda==4.5.10
configobj==5.0.6
configparser==3.5.0
contextlib2==0.5.4
cryptography==1.7.1
cx-Freeze==5.0.2
cycler==0.10.0
Cython==0.28.1
cytoolz==0.8.2
dask==0.17.4
#datashape==0.5.4
decorator==4.0.11
dill==0.2.5
docutils==0.14
EasyProcess==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
future==0.16.0
gast==0.2.0
gensim==3.4.0
gevent==1.2.1
glove-python==0.1.0
graphviz==0.8.1
greenlet==0.4.11
grpcio==1.14.1
gym==0.9.6
h5py==2.7.1
HeapDict==1.0.0
hgtk==0.1.1
html5lib==0.9999999
idna==2.6
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.9.4
jmespath==0.9.3
JPype1==0.6.2
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==2.1.2
keras-rl==0.4.2
kiwisolver==1.0.1
konlpy==0.4.4
lazy-object-proxy==1.2.2
llvmlite==0.21.0
locket==0.2.0
lxml==3.7.2
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
#menuinst==1.4.4
mistune==0.7.3
mpmath==0.19
multidict==3.3.2
multipledispatch==0.4.9
mxnet==0.12.1
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.5
nose==1.3.7
notebook==4.3.1
numba==0.36.2
numexpr==2.6.2
numpy==1.14.5
numpydoc==0.6.0
odo==0.5.0
olefile==0.45.1
openpyxl==2.4.1
#pandas==0.22.0
partd==0.3.7
#path.py==0.0.0
pathlib2==2.2.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
ply==3.9
prompt-toolkit==1.0.9
protobuf==3.6.1
psutil==5.0.1
py==1.4.32
pyasn1==0.1.9
pycares==2.3.0
pychrome==0.2.2
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0.2
pyduktape==0.0.6
pyflakes==1.5.0
pyglet==1.3.0
Pygments==2.1.3
PyJWT==1.6.4
pylint==1.6.4
pymongo==3.6.0
pyOpenSSL==16.2.0
pyparsing==2.1.4
pyperclip==1.6.0
pytest==3.0.5
python-dateutil==2.6.1
pytz==2016.10
PyVirtualDisplay==0.2.1
pywin32==220
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
QtPy==1.2.1
regex==2018.2.21
requests==2.18.4
requests-jwt==0.4
rope-py3k==0.9.4.post1
s3transfer==0.1.13
scikit-learn==0.19.0
scipy==1.0.1
seaborn==0.8.1
selenium==3.4.3
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.7
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sox==1.3.0
sphinx==1.5.1
SQLAlchemy==1.1.5
statsmodels==0.8.0
sympy==1.0
TA-Lib==0.4.17
tables==3.4.2
tailer==0.4.1
tensorboard==1.10.0
tensorflow-gpu==1.10.0
tensorflow-tensorboard==0.4.0rc3
tensorforce==0.4.3
termcolor==1.1.0
toolz==0.8.2
tornado==4.4.2
tqdm==4.19.4
traitlets==4.3.1
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
Werkzeug==0.14.1
widgetsnbextension==1.2.6
win-unicode-console==0.5
wincertstore==0.2
wmctrl==0.3
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.2
xlwt==1.2.0
yarl==0.17.0
ZigZag==0.2.2
